+++ CHANGES.rst
@@ -11,6 +11,10 @@ Compute
   (GITHUB-1628)
   [Arturo Noha - @r2ronoha, Tomaz Muraus - @Kami]
 
+- [OpenStack] Add Server Groups functions in OpenStack driver.
+  (GITHUB-1629)
+  [Miguel Caballer - @micafer]
+
 Storage
 ~~~~~~~